Removes legacy APNS replaced with v2
The binary interface was shut down by Apple. Fixes #568.
