+++ README.md
@@ -55,12 +55,11 @@ $ bundle exec rpush init
 
 There is a choice of two modes (and one legacy mode) using certificates or using tokens:
 
-* `Rpush::Apns2` This requires an annually renewable certificate. see https://developer.apple.com/documentation/usernotifications/setting_up_a_remote_notification_server/establishing_a_certificate-based_connection_to_apns
+* `Rpush::Apns` This requires an annually renewable certificate. see https://developer.apple.com/documentation/usernotifications/setting_up_a_remote_notification_server/establishing_a_certificate-based_connection_to_apns
 * `Rpush::Apnsp8` This uses encrypted tokens and requires an encryption key id and encryption key (provide as a p8 file). (see https://developer.apple.com/documentation/usernotifications/setting_up_a_remote_notification_server/establishing_a_token-based_connection_to_apns)
-* `Rpush::Apns` There is also the original APNS (the original version using certificates with a binary underlying protocol over TCP directly rather than over Http/2).
   Apple have [announced](https://developer.apple.com/news/?id=c88acm2b) that this is not supported after March 31, 2021.
 
-If this is your first time using the APNs, you will need to generate either SSL certificates (for Apns2 or Apns) or an Encryption Key (p8) and an Encryption Key ID (for Apnsp8). See [Generating Certificates](https://github.com/rpush/rpush/wiki/Generating-Certificates) for instructions.
+If this is your first time using the APNs, you will need to generate either SSL certificates (for standard Apns) or an Encryption Key (p8) and an Encryption Key ID (for Apnsp8). See [Generating Certificates](https://github.com/rpush/rpush/wiki/Generating-Certificates) for instructions.
 
 ##### Apnsp8
 
@@ -87,12 +86,12 @@ n.data = { foo: :bar }
 n.save!
 ```
 
-##### Apns2
+##### Apns
 
 (NB this uses the same protocol as Apnsp8, but authenticates with a certificate rather than tokens)
 
 ```ruby
-app = Rpush::Apns2::App.new
+app = Rpush::Apns::App.new
 app.name = "ios_app"
 app.certificate = File.read("/path/to/sandbox.pem")
 app.environment = "development"
@@ -103,8 +102,8 @@ app.save!
 ```
 
 ```ruby
-n = Rpush::Apns2::Notification.new
-n.app = Rpush::Apns2::App.find_by_name("ios_app")
+n = Rpush::Apns::Notification.new
+n.app = Rpush::Apns::App.find_by_name("ios_app")
 n.device_token = "..." # hex string
 n.alert = "hi mom!"
 n.data = {
@@ -116,34 +115,13 @@ n.save!
 
 You should also implement the [ssl_certificate_will_expire](https://github.com/rpush/rpush/wiki/Reflection-API) reflection to monitor when your certificate is due to expire.
 
-##### Apns (legacy protocol)
-
-```ruby
-app = Rpush::Apns::App.new
-app.name = "ios_app"
-app.certificate = File.read("/path/to/sandbox.pem")
-app.environment = "development" # APNs environment.
-app.password = "certificate password"
-app.connections = 1
-app.save!
-```
-
-```ruby
-n = Rpush::Apns::Notification.new
-n.app = Rpush::Apns::App.find_by_name("ios_app")
-n.device_token = "..." # hex string
-n.alert = "hi mom!"
-n.data = { foo: :bar }
-n.save!
-```
-
 ##### Safari Push Notifications
 
 Using one of the notifications methods above, the `url_args` attribute is available for Safari Push Notifications.
 
 ##### Environment
 
-The app `environment` for any Apns* option is "development" for XCode installs, and "production" for app store and TestFlight. Note that for Apns2 you can now use one (production + sandbox) certificate (you don't need a separate "sandbox" or development certificate), but if you do generate a development/sandbox certificate it can only be used for "development". With Apnsp8 tokens, you can target either "development" or "production" environments.
+The app `environment` for any Apns* option is "development" for XCode installs, and "production" for app store and TestFlight. Note that you can now use one (production + sandbox) certificate (you don't need a separate "sandbox" or development certificate), but if you do generate a development/sandbox certificate it can only be used for "development". With Apnsp8 tokens, you can target either "development" or "production" environments.
 
 #### Firebase Cloud Messaging
