+++ lib/rpush/client/active_record.rb
@@ -10,9 +10,6 @@ require 'rpush/client/active_record/apns/notification'
 require 'rpush/client/active_record/apns/feedback'
 require 'rpush/client/active_record/apns/app'
 
-require 'rpush/client/active_record/apns2/notification'
-require 'rpush/client/active_record/apns2/app'
-
 require 'rpush/client/active_record/apnsp8/notification'
 require 'rpush/client/active_record/apnsp8/app'
