--- lib/rpush/client/active_record/apns2/notification.rb
@@ -1,13 +1,0 @@
-module Rpush
-  module Client
-    module ActiveRecord
-      module Apns2
-        class Notification < Rpush::Client::ActiveRecord::Notification
-          include Rpush::Client::ActiveModel::Apns::Notification
-          include Rpush::Client::ActiveModel::Apns2::Notification
-          include Rpush::Client::ActiveRecord::Apns::ActiveRecordSerializableNotification
-        end
-      end
-    end
-  end
-end
