+++ lib/rpush/client/active_record/apnsp8/notification.rb
@@ -3,7 +3,7 @@ module Rpush
     module ActiveRecord
       module Apnsp8
         class Notification < Rpush::Client::ActiveRecord::Apns::Notification
-          include Rpush::Client::ActiveModel::Apns2::Notification
+          include Rpush::Client::ActiveModel::Apns::Notification
         end
       end
     end
