+++ lib/rpush/client/redis.rb
@@ -21,9 +21,6 @@ require 'rpush/client/redis/apns/app'
 require 'rpush/client/redis/apns/notification'
 require 'rpush/client/redis/apns/feedback'
 
-require 'rpush/client/redis/apns2/app'
-require 'rpush/client/redis/apns2/notification'
-
 require 'rpush/client/redis/apnsp8/app'
 require 'rpush/client/redis/apnsp8/notification'
