+++ lib/rpush/client/redis/apnsp8/notification.rb
@@ -4,7 +4,6 @@ module Rpush
       module Apnsp8
         class Notification < Rpush::Client::Redis::Notification
           include Rpush::Client::ActiveModel::Apns::Notification
-          include Rpush::Client::ActiveModel::Apns2::Notification
           include Rpush::Client::ActiveModel::Apnsp8::Notification
         end
       end
