--- lib/rpush/daemon/apns2.rb
@@ -1,10 +1,0 @@
-module Rpush
-  module Daemon
-    module Apns2
-      extend ServiceConfigMethods
-
-      batch_deliveries true
-      dispatcher :apns_http2
-    end
-  end
-end
