+++ spec/functional/apns_spec.rb
@@ -1,24 +1,48 @@
 require 'functional_spec_helper'
 
-describe 'APNs' do
+describe 'APNs http2 adapter' do
+  let(:fake_client) {
+    double(
+      prepare_request: fake_http2_request,
+      close:           'ok',
+      call_async:      'ok',
+      join:            'ok',
+      on:              'ok'
+    )
+  }
   let(:app) { create_app }
-  let(:tcp_socket) { double(TCPSocket, setsockopt: nil, close: nil) }
-  let(:ssl_socket) { double(OpenSSL::SSL::SSLSocket, :sync= => nil, connect: nil, write: nil, flush: nil, read: nil, close: nil) }
-  let(:io_double) { double(select: nil) }
-  let(:delivered_ids) { [] }
-  let(:failed_ids) { [] }
-  let(:retry_ids) { [] }
+  let(:fake_device_token) { 'a' * 108 }
+  let(:fake_http2_request) { double }
+  let(:fake_http_resp_headers) {
+    {
+      ":status" => "200",
+      "apns-id"=>"C6D65840-5E3F-785A-4D91-B97D305C12F6"
+    }
+  }
+  let(:fake_http_resp_body) { '' }
+  let(:notification_data) { nil }
 
   before do
     Rpush.config.push_poll = 0.5
-    stub_tcp_connection(tcp_socket, ssl_socket, io_double)
+    allow(NetHttp2::Client).
+      to receive(:new).and_return(fake_client)
+    allow(fake_http2_request).
+      to receive(:on).with(:headers).
+      and_yield(fake_http_resp_headers)
+    allow(fake_http2_request).
+      to receive(:on).with(:body_chunk).
+      and_yield(fake_http_resp_body)
+    allow(fake_http2_request).
+      to receive(:on).with(:close).
+      and_yield
   end
 
   def create_app
     app = Rpush::Apns::App.new
     app.certificate = TEST_CERT
     app.name = 'test'
-    app.environment = 'sandbox'
+    app.environment = 'development'
+    app.bundle_id = 'com.example.app'
     app.save!
     app
   end
@@ -26,136 +50,279 @@ describe 'APNs' do
   def create_notification
     notification = Rpush::Apns::Notification.new
     notification.app = app
+    notification.sound = 'default'
     notification.alert = 'test'
-    notification.device_token = 'a' * 108
+    notification.device_token = fake_device_token
+    notification.data = notification_data
+    notification.content_available = 1
     notification.save!
     notification
   end
 
-  def wait
-    sleep 0.1
-  end
-
-  def wait_for_notification_to_deliver(notification)
-    timeout { wait until delivered_ids.include?(notification.id) }
-  end
-
-  def wait_for_notification_to_fail(notification)
-    timeout { wait until failed_ids.include?(notification.id) }
-  end
-
-  def wait_for_notification_to_retry(notification)
-    timeout { wait until retry_ids.include?(notification.id) }
-  end
-
-  def fail_notification(notification)
-    allow(ssl_socket).to receive_messages(read: [8, 4, notification.id].pack('ccN'))
-    enable_io_select
-  end
-
-  def enable_io_select
-    called = false
-    allow(io_double).to receive(:select) do
-      if called
-        nil
-      else
-        called = true
-      end
-    end
-  end
-
   it 'delivers a notification successfully' do
     notification = create_notification
+
+    thread = nil
+    expect(fake_http2_request).
+      to receive(:on).with(:close) { |&block|
+        # imitate HTTP2 delay
+        thread = Thread.new { sleep(0.01); block.call }
+      }
+    expect(fake_client).to receive(:join) { thread.join }
+
+    expect(fake_client)
+      .to receive(:prepare_request)
+      .with(
+        :post,
+        "/3/device/#{fake_device_token}",
+        { body: "{\"aps\":{\"alert\":\"test\",\"sound\":\"default\",\"content-available\":1}}",
+          headers: {
+            'apns-expiration' => '0',
+            'apns-priority' => '10',
+            'apns-topic' => 'com.example.app',
+            'apns-push-type' => 'background'
+          }
+        }
+      )
+      .and_return(fake_http2_request)
+
     expect do
       Rpush.push
       notification.reload
     end.to change(notification, :delivered).to(true)
   end
 
-  it 'receives feedback' do
-    app
-    tuple = "N\xE3\x84\r\x00 \x83OxfU\xEB\x9F\x84aJ\x05\xAD}\x00\xAF1\xE5\xCF\xE9:\xC3\xEA\a\x8F\x1D\xA4M*N\xB0\xCE\x17"
-    allow(ssl_socket).to receive(:read).and_return(tuple, nil)
-    Rpush.apns_feedback
-    feedback = Rpush::Apns::Feedback.all.first
-    expect(feedback).not_to be_nil
-    expect(feedback.app_id).to eq(app.id)
-    expect(feedback.device_token).to eq('834f786655eb9f84614a05ad7d00af31e5cfe93ac3ea078f1da44d2a4eb0ce17')
+  context 'when there is "headers" field in a data' do
+    let(:bundle_id) { 'some.example.com' }
+    let(:notification_data) {
+      {
+        'headers' => { 'apns-topic' => bundle_id },
+        'some_field' =>  'some value'
+      }
+    }
+
+    it 'delivers notification with custom headers' do
+      notification = create_notification
+
+      expect(fake_client)
+        .to receive(:prepare_request)
+        .with(
+          :post,
+          "/3/device/#{fake_device_token}",
+          { body: "{\"aps\":{\"alert\":\"test\",\"sound\":\"default\","\
+                  "\"content-available\":1},\"some_field\":\"some value\"}",
+            headers: {
+              'apns-topic' => bundle_id,
+              'apns-expiration' => '0',
+              'apns-priority' => '10',
+              'apns-push-type' => 'background'
+            }
+          }
+        ).and_return(fake_http2_request)
+
+      expect do
+        Rpush.push
+        notification.reload
+      end.to change(notification, :delivered).to(true)
+    end
   end
 
   describe 'delivery failures' do
-    before do
-      Rpush.reflect do |on|
-        on.notification_delivered do |n|
-          delivered_ids << n.id
-        end
+    context 'when response is about incorrect request' do
+      let(:fake_http_resp_headers) {
+        {
+          ":status" => "404",
+          "apns-id"=>"C6D65840-5E3F-785A-4D91-B97D305C12F6"
+        }
+      }
 
-        on.notification_id_failed do |_, n_id|
-          failed_ids << n_id
-        end
-
-        on.notification_id_will_retry do |_, n_id|
-          retry_ids << n_id
-        end
-
-        on.notification_will_retry do |n|
-          retry_ids << n.id
-        end
+      it 'fails to deliver a notification' do
+        notification = create_notification
+        expect do
+          Rpush.push
+          notification.reload
+        end.to change(notification, :failed).to(true)
       end
 
-      Rpush.embed
-    end
+      it 'reflects :notification_id_failed' do
+        Rpush.reflect do |on|
+          on.notification_id_failed do |app, id, code, descr|
+            expect(app).to be_kind_of(Rpush::Client::Apns::App)
+            expect(id).to eq 1
+            expect(code).to eq 404
+            expect(descr).to be_nil
+          end
+        end
 
-    after do
-      Rpush.reflection_stack.clear
-      Rpush.reflection_stack.push(Rpush::ReflectionCollection.new)
-
-      timeout { Rpush.shutdown }
-    end
-
-    it 'fails to deliver a notification' do
-      notification = create_notification
-      wait_for_notification_to_deliver(notification)
-      fail_notification(notification)
-      wait_for_notification_to_fail(notification)
-    end
-
-    describe 'with a failed connection' do
-      it 'retries all notifications' do
-        allow_any_instance_of(Rpush::Daemon::TcpConnection).to receive_messages(sleep: nil)
-        expect(ssl_socket).to receive(:write).at_least(1).times.and_raise(Errno::EPIPE)
-        notifications = 2.times.map { create_notification }
-        notifications.each { |n| wait_for_notification_to_retry(n) }
+        notification = create_notification
+        Rpush.push
       end
     end
 
-    describe 'with multiple notifications' do
-      let(:notification1) { create_notification }
-      let(:notification2) { create_notification }
-      let(:notification3) { create_notification }
-      let(:notification4) { create_notification }
-      let(:notifications) { [notification1, notification2, notification3, notification4] }
+    context 'when response returns 500 error for APNs maintenance' do
+      let(:fake_http_resp_headers) {
+        {
+          ":status" => "500",
+          "apns-id"=>"C6D65840-5E3F-785A-4D91-B97D305C12F6"
+        }
+      }
 
-      it 'marks the correct notification as failed' do
-        notifications.each { |n| wait_for_notification_to_deliver(n) }
-        fail_notification(notification2)
-        wait_for_notification_to_fail(notification2)
+      it 'fails but retries delivery several times' do
+        notification = create_notification
+        expect do
+          Rpush.push
+          notification.reload
+        end.to change(notification, :retries)
       end
 
-      it 'does not mark prior notifications as failed' do
-        notifications.each { |n| wait_for_notification_to_deliver(n) }
-        fail_notification(notification2)
-        wait_for_notification_to_fail(notification2)
+      it 'reflects :notification_id_will_retry' do
+        Rpush.reflect do |on|
+          on.notification_id_will_retry do |app, id, timer|
+            expect(app).to be_kind_of(Rpush::Client::Apns::App)
+            expect(id).to eq 1
+          end
+        end
 
-        expect(failed_ids).to_not include(notification1.id)
-        notification1.reload
-        expect(notification1.delivered).to eq(true)
+        notification = create_notification
+        Rpush.push
+      end
+    end
+
+    context 'when there is SocketError' do
+      let(:fake_http_resp_headers) {
+        {
+          ":status" => "500",
+          "apns-id"=>"C6D65840-5E3F-785A-4D91-B97D305C12F6"
+        }
+      }
+
+      before(:each) do
+        expect(fake_client).to receive(:call_async) { raise(SocketError) }
       end
 
-      it 'marks notifications following the failed one as retryable' do
-        notifications.each { |n| wait_for_notification_to_deliver(n) }
-        fail_notification(notification2)
-        [notification3, notification4].each { |n| wait_for_notification_to_retry(n) }
+      it 'fails but retries delivery several times' do
+        notification = create_notification
+        expect do
+          Rpush.push
+          notification.reload
+        end.to change(notification, :retries)
+      end
+
+      it 'reflects :notification_id_will_retry' do
+        Rpush.reflect do |on|
+          on.notification_id_will_retry do |app, id, timer|
+            expect(app).to be_kind_of(Rpush::Client::Apns::App)
+            expect(id).to eq 1
+            expect(timer).to be_kind_of(Time)
+          end
+        end
+
+        notification = create_notification
+        Rpush.push
+      end
+
+      context 'when specific notification was delivered before request failed' do
+        let(:fake_http_resp_headers) {
+          {
+            ":status" => "200",
+            "apns-id"=>"C6D65840-5E3F-785A-4D91-B97D305C12F6"
+          }
+        }
+
+        it 'fails but will not retry this notification' do
+          notification = create_notification
+          expect do
+            Rpush.push
+            notification.reload
+          end.to change(notification, :retries).by(0)
+             .and change(notification, :delivered).to(true)
+        end
+      end
+    end
+
+    context 'when any StandardError occurs' do
+      before(:each) do
+        expect(fake_client).to receive(:call_async) { raise(StandardError) }
+      end
+
+      it 'marks notification failed' do
+        notification = create_notification
+        expect do
+          Rpush.push
+          notification.reload
+        end.to change(notification, :failed).to(true)
+      end
+
+      it 'reflects :error' do
+        Rpush.reflect do |on|
+          on.error do |error|
+            expect(error).to be_kind_of(StandardError)
+           reflector.accept
+          end
+        end
+
+        notification = create_notification
+        Rpush.push
+      end
+    end
+
+    context 'when waiting for requests to complete times out' do
+      let(:on_close) do
+        proc { |&block| @thread = Thread.new { sleep(0.01) } }
+      end
+
+      before(:each) do
+        @thread = nil
+
+        expect(fake_http2_request).
+          to receive(:on).with(:close), &on_close
+
+        expect(fake_client).to receive(:join) { @thread.join; raise(NetHttp2::AsyncRequestTimeout) }
+      end
+
+      it 'closes the client' do
+        create_notification
+        expect(fake_client).to receive(:close)
+        Rpush.push
+      end
+
+      it 'reflects :error' do
+        reflected_error = false
+        Rpush.reflect do |on|
+          on.error do |error|
+            reflected_error = true
+            expect(error).to be_kind_of(StandardError)
+            reflector.accept
+          end
+        end
+
+        notification = create_notification
+        Rpush.push
+
+        expect(reflected_error).to be true
+      end
+
+      it 'fails but retries delivery several times' do
+        notification = create_notification
+        expect do
+          Rpush.push
+          notification.reload
+        end.to change(notification, :retries)
+      end
+
+      context 'when specific notification was delivered before another async call failed' do
+        let(:on_close) do
+          proc { |&block| @thread = Thread.new { sleep(0.01); block.call } }
+        end
+
+        it 'fails but retries delivery several times' do
+          notification = create_notification
+          expect do
+            Rpush.push
+            notification.reload
+          end.to change(notification, :retries).by(0)
+             .and change(notification, :delivered).to(true)
+        end
       end
     end
   end
