+++ spec/unit/client/active_record/apns/notification_spec.rb
@@ -8,17 +8,17 @@ describe Rpush::Client::ActiveRecord::Apns::Notification do
 
   it "should validate the length of the binary conversion of the notification" do
     notification = described_class.new
-    notification.app = Rpush::Apns2::App.create(name: 'test', environment: 'development', certificate: TEST_CERT)
+    notification.app = Rpush::Apns::App.create(name: 'test', environment: 'development', certificate: TEST_CERT)
     notification.device_token = "a" * 108
     notification.alert = ""
 
-    notification.alert << "a" until notification.payload.bytesize == 2048
+    notification.alert << "a" until notification.payload.bytesize == 4096
     expect(notification.valid?).to be_truthy
     expect(notification.errors[:base]).to be_empty
 
     notification.alert << "a"
     expect(notification.valid?).to be_falsey
-    expect(notification.errors[:base].include?("APN notification cannot be larger than 2048 bytes. Try condensing your alert and device attributes.")).to be_truthy
+    expect(notification.errors[:base].include?("APN notification cannot be larger than 4096 bytes. Try condensing your alert and device attributes.")).to be_truthy
   end
 
   describe "multi_json usage" do
