--- spec/unit/client/active_record/apns2/app_spec.rb
@@ -1,5 +1,0 @@
-require 'unit_spec_helper'
-
-describe Rpush::Client::ActiveRecord::Apns2::App do
-  it_behaves_like 'Rpush::Client::Apns::App'
-end if active_record?
