--- spec/unit/client/redis/apns2/app_spec.rb
@@ -1,4 +1,0 @@
-require 'unit_spec_helper'
-
-describe Rpush::Client::Redis::Apns2::App do
-end if redis?
